Updated DiveTableTest class (Still working on it)
Working on the DiveTableTest class. I have copied the test from the
DiveTable class and put it in this class; I was able to enhance that
just a bit. I’m working on figuring out how to test the
calculations/functions directly. Will try to get a more complete test
out later this week.