Updates to Test Code, but still working on it
Was able to add a few more tests. With the 3rd test, I got a
NullPointerException when I added a second dive after the first dive.
The first dive had a depth of 100, time of 30, then SIT of 20. The
second dive had a depth of 40, time of 75, and SIT of 75. Will work on
testing it more.