Updated DiveTableTest. Slight change in DiveTable.
Added more tests to DiveTableTest class. For endOfDiveTable and
surfaceIntervalTimeTable, I was able to do extensive testing and find
no errors. For the repetitiveDiveTimeTable, I wasn’t able to test it
that much. Only did test to make sure it had two values in arrays. From
there, code was manually checked and found small discrepancy in
repetitiveDiveTimeTable. In Row J, Column 50, the AMDT was supposed to
be 0 and not 4. That was changed once it was detected.